requirements.txt
folium==0.14.0
websockets==12.0
python-dotenv==1.0.0
pytest==7.4.3